requirements.txt
trafilatura>=1.6.0
spacy>=3.7.0
scikit-learn>=1.3.0
textstat>=0.7.3
nltk>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.39.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
lxml>=4.9.0
python-dotenv>=1.0.0
openai>=1.0.0
wordcloud>=1.9.0
plotly>=5.18.0
matplotlib>=3.8.0
